% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex-utils.R
\name{bfx_fetch}
\alias{bfx_fetch}
\title{Get Bitfenix POST Headers}
\usage{
bfx_fetch(body)
}
\arguments{
\item{body}{a \code{list} containing the POST body.}
}
\value{
A \code{list} containing POST headers. See details.
}
\description{
Get Bitfenix POST Headers
}
\details{
\itemize{
\item \code{api_key}: User's Bitfenix API key.
\item \code{nonce}: The cryptographic nonce used. Generated via: (System time * 1000)
\item \code{body}: The JSON stringify'd POST body passed to the function.
\item \code{signature}: The authentication signature hashed with your API secret key.
}
}
\keyword{internal}
