% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex.R
\name{get_status}
\alias{get_status}
\title{Get Bitfinex platform status}
\usage{
get_status()
}
\value{
Either "Operative" if platform is usable, or "Maintenace".
}
\description{
Calls the Bitfinex API to determine platform operative status.
}
\examples{
\dontrun{
    print(get_status())
    # Output:
    # [1] "Operative"
}
}
