% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex-utils.R
\name{set_names}
\alias{set_names}
\title{Set ticker tibble colnames (Bitfenix)}
\usage{
set_names(ticker_info_get)
}
\description{
See \code{\link{ticker}} and \code{\link{tickers}}.
}
\keyword{internal}
