% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Get API status for supported platforms}
\usage{
status()
}
\value{
Console output describing the status of each
supported API with either "GOOD" or "BAD".
}
\description{
Get API status for supported platforms
}
\examples{
\dontrun{
    cryptocurr::status()

    # OUTPUT:
    # cryptocurr
    # ===============
    # BITFENIX GOOD
    # COINBASE BAD
    # KRAKEN   BAD
    # BINANCE  BAD
}
}
