% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex-public.R
\name{ticker}
\alias{ticker}
\title{Retrieve Cryptocurrency Data}
\usage{
ticker(SYMBOL = "tBTCUSD")
}
\arguments{
\item{SYMBOL}{Character ticker symbol for a single
cryptocurrency. Must be prepended with a \code{t} or
\code{f} to specify trading or funding
cryptocurrencies, respectively.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
See details for column names.
}
\description{
Calls the Bitfinex API to retrieve cryptocurrency data
for a single ticker.
}
\details{
\subsection{\code{tibble} descriptions:}{
\itemize{
\item \code{SYMBOL}: The symbol/ticker of the requested data
\item \code{FRR}: Flash Return Rate - average of all fixed rate funding over
the last hour \emph{(funding tickers only)}
\item \code{BID}: Price of last highest bid
\item \code{BID_PERIOD}: Bid period covered in days \emph{(funding tickers only)}
\item \code{BID_SIZE}: Sum of the 25 highest bid sizes
\item \code{ASK}: Price of last lowest ask
\item \code{ASK_PERIOD}: Ask period covered in days \emph{(funding tickers only)}
\item \code{ASK_SIZE}: Sum of the 25 lowest ask sizes
\item \code{DAILY_CHANGE}: Amount that the last price has changed since yesterday
\item \code{DAILY_CHANGE_RELATIVE}: Relative price change since yesterday
(*100 for percentage change)
\item \code{LAST_PRICE}: Price of the last trade
\item \code{VOLUME}: Daily volume
\item \code{HIGH}: Daily high
\item \code{LOW}: Daily low
\item \code{FRR_AMOUNT_AVAILABLE}: The amount of funding that is available at
the Flash Return Rate \emph{(funding tickers only)}
}
}
}
\examples{
\dontrun{
    # Get Bitcoin info
    ticker(SYMBOL = "tBTCUSD")

    # Get USD info
    tickers(SYMBOL = "USD")
}
}
