% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex-public.R
\name{ticker_history}
\alias{ticker_history}
\title{Retrieve Cryptocurrency Historical Data}
\usage{
ticker_history(SYMBOLS = c("ALL"), start, end, limit)
}
\arguments{
\item{SYMBOLS}{Character vector containing ticker symbols for
cryptocurrencies, or \code{ALL} for all possible tickers.
Must be prepended with a \code{t} or \code{f} to specify
trading or funding cryptocurrencies, respectively.}

\item{start}{Millisecond start time}

\item{end}{Millisecond end time}

\item{limit}{Number of records (Max 250)}
}
\value{
A \link[tibble:tibble-package]{tibble}. See details for column names.
}
\description{
Calls the Bitfinex API to retrieve historical cryptocurrency
records.
}
\details{
\subsection{\code{tibble} descriptions:}{
\itemize{
\item \code{SYMBOL}: The symbol/ticker of the requested data
\item \code{BID}: Price of last highest bid
\item \code{ASK}: Price of last lowest ask
\item \code{MTS}: millisecond timestamp
}
}
}
\examples{
\dontrun{
    # Get Bitcoin and Litecoin info
    tickers(SYMBOLS = c("tBTCUSD", "tLTCUSD"))
    # Returns $trading tibble with 2 rows
    # and $funding tibble with 0 rows

    # Get USD info
    tickers(SYMBOLS = "USD")
    # Return $trading tibble with 0 rows
    # and $funding tibble with 1 row
}
}
