% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex-public.R
\name{trades}
\alias{trades}
\title{Retrieve Cryptocurrency Trade Data}
\usage{
trades(SYMBOL = "tBTCUSD", start, end, limit)
}
\arguments{
\item{SYMBOL}{Character vector containing ticker symbols for
cryptocurrencies.
Must be prepended with a \code{t} or \code{f} to specify
trading or funding cryptocurrencies, respectively.}

\item{start}{Millisecond start time}

\item{end}{Millisecond end time}

\item{limit}{Number of records (Max 10000)}
}
\value{
A \link[tibble:tibble-package]{tibble}. See details for column names.
}
\description{
Calls the Bitfinex API to retrieve cryptocurrency public trade
records.
}
\details{
\subsection{\code{tibble} descriptions:}{
\itemize{
\item \code{SYMBOL}: The symbol/ticker of the requested data
\item \code{ID}: ID of the trade
\item \code{MTS}: millisecond timestamp
\item \code{AMOUNT}: 	How much was bought (positive) or sold (negative).
\item \code{PRICE}: Price at which the trade was executed (trading tickers only)
\item \code{RATE}: Rate at which funding transaction occurred (funding tickers only)
\item \code{PERIOD}: Amount of time the funding transaction was for
(funding tickers only)
}
}
}
\examples{
\dontrun{
    # Get 1000 Bitcoin trades info
    trades(SYMBOL = "tBTCUSD", limit = 1000)

    # Get USD trades info
    trades(SYMBOL = "USD")
}
}
