% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex-utils.R
\name{set_bfx_auth}
\alias{set_bfx_auth}
\title{Set Bitfenix API Keys}
\usage{
set_bfx_auth(API_KEY, API_SECRET)
}
\arguments{
\item{API_KEY}{Bitfenix API key}

\item{API_SECRET}{Bitfenix API Secret key}
}
\value{
\code{TRUE} if keys were set.
}
\description{
Set Bitfenix API Keys
}
